+++ bin/cli.js
@@ -1,186 +1,164 @@
 #!/usr/bin/env node
 
-/* global __dirname */
-/* global process */
-
-var nokit = require("../");
-var path = require("path");
-var fs = require("fs");
-var domain = require("domain");
-var Notifier = require('./notifier');
-var processLog = require('./process-log');
-var processMgr = require('./process-mgr');
-var CmdLine = require('cmdline');
-var bootstrap = require('./bootstrap');
-var utils = nokit.utils;
-var pkg = nokit.pkg;
-var console = nokit.console;
-var env = nokit.env;
+const nokit = require("../");
+const path = require("path");
+const fs = require("fs");
+const domain = require("domain");
+const Notifier = require('./notifier');
+const processLog = require('./process-log');
+const processMgr = require('./process-mgr');
+const cmdline = require('cmdline');
+const startup = require('./startup');
+const utils = nokit.utils;
+const pkg = nokit.pkg;
+const console = nokit.console;
+const env = nokit.env;
 //var _debugger = require('../tool/debugger');
 
-//工作目录
-var cwd = process.cwd();
-
-/**
- * 输出帮助信息
- **/
-function printInfo(versionOnly) {
-  console.log(pkg.displayName + " " + pkg.version, true);
-  if (!versionOnly) {
-    console.log(env.EOL + "Usage:", true);
-    console.log("  nokit <command> [params] [options]" + env.EOL, true);
-    console.log("Commands:", true);
-    console.log("  create     [name] [mvc|nsp|rest] [folder]", true);
-    console.log("  start      [port] [root]", true);
-    console.log("             [-name:<name>]", true);
-    console.log("             [-env:<name>]", true);
-    console.log("             [-cluster[:num]]", true);
-    console.log("             [-watch[:.ext,...]]", true);
-    console.log("             [node-opts]", true);
-    console.log("  stop       [name|pid|all]", true);
-    console.log("  restart    [name|pid|all]", true);
-    console.log("  list       (no params)", true);
-    console.log("  delete     [name|pid|all]", true);
-    console.log("  bootstrap  [true|false]" + env.EOL, true);
-  }
-}
-
-var dm = domain.create();
+const dm = domain.create();
 dm.on('error', function (err) {
   console.error(err.message + env.EOL + err.stack);
 });
 
 dm.run(function () {
 
-  var notifier = new Notifier();
+  const notifier = new Notifier();
 
-  var cml = new CmdLine({
-    commandEnabled: true
-  });
+  cmdline
+    .error(function (err) {
+      console.error(err.message);
+    })
+    .version(pkg.displayName + " " + pkg.version)
+    .help(path.normalize('@' + __dirname + '/help.txt'))
 
-  switch (cml.command) {
-    case "":
-      printInfo(cml.options.has('-v'));
-      break;
-    case "?":
-    case "help":
-      printInfo();
-      break;
-    case "new":
-    case "create":
+    /**
+     * 创建一个应用
+     **/
+    .root.command(['new', 'create'])
+    .option(['-t', '--type'], 'string')
+    .action(function ($1, $2, type) {
       console.log("Creating...");
       //处理参数
-      var appName = cml.args[0] || 'nokit-app';
-      var appType = cml.args[1] || 'mvc';
-      var dstPath = cml.args[2] || "./";
+      var appName = $1 || 'nokit-app';
+      var dstPath = $2 || "./";
+      var appType = type || 'mvc';
       //处理路径
-      var dstFullPath = path.resolve(cwd, path.normalize(dstPath + '/' + appName));
+      var dstFullPath = path.resolve(process.cwd(), path.normalize(dstPath + '/' + appName));
       var srcFullPath = path.resolve(__dirname, path.normalize('../examples/' + appType));
       //检查目标路径是否已存在
       if (fs.existsSync(dstFullPath)) {
-        console.error('Create failure, directory "' + dstFullPath + '" already exists');
+        console.error('Create failed, folder "' + dstFullPath + '" already exists');
       } else {
         //复制应用模板
         nokit.utils.copyDir(srcFullPath, dstFullPath);
         console.log('In the "' + dstFullPath + '" has been created');
       }
-      break;
-    case "start":
+      return false;
+    }, false)
+
+    /**
+     * 启动一个应用
+     **/
+    .root.command(['start'])
+    .option(['-n', '--name'], 'string')
+    .option(['-p', '--port'], 'number')
+    .option(['-c', '--cluster'], 'number')
+    .option(['-w', '--watch'], 'switch')
+    .option(['-e', '--env'], 'string')
+    .option(['--config'], 'string')
+    .option(['--public'], 'string')
+    .option(['--node'], 'string*')
+    .action(function () {
       console.log("Starting...");
-      var startInfo = [];
-      //处理调式参数
-      //因为 master 使用 --debug 指定的端口，而 worker 的 debug port 则会每个 +1
-      //所以这里，把 --debug 的端口 -1 作为 master 的 port 
-      //这样第一个 worker 的端口就是 --debug 指定的端口了，后续 worker 会每个 +1
-      var debugPort = null;
-      if (cml.options.has('--debug')) {
-        debugPort = parseInt(cml.options.getValue('--debug') || 5858) - 1;
-        cml.options.setValue('--debug', debugPort);
-      }
-      if (cml.options.has('--debug-brk')) {
-        debugPort = parseInt(cml.options.getValue('--debug-brk') || 5858) - 1;
-        cml.options.setValue('--debug-brk', debugPort);
-      }
-      //添加 node 控制选项 
-      var nodeOptions = cml.options.getNodeOptions();
-      nodeOptions.forEach(function (item) {
-        startInfo.push(item);
-      });
+      //定义 params
+      var params = {};
       //添加入口程序
-      var appFile = path.normalize(__dirname + '/app.js');
-      startInfo.push(appFile);
-      //添加监听端口
-      var port = cml.args[0] || 0;
-      startInfo.push(port);
+      params.main = path.normalize(__dirname + '/app.js');
       //添加应用根目录
-      var root = path.resolve(cwd, path.normalize(cml.args[1] || './'));
-      startInfo.push(root);
+      params.root = path.resolve(process.cwd(), path.normalize(this.argv[0] || './'));
+      //添加监听端口
+      params.port = this.get('port');
       //添加控制选项
-      cml.options.forEach(function (item) {
-        startInfo.push(item);
+      utils.each(this.options, function (name, value) {
+        params[name] = value;
       });
-      //console.log("startInfo: " + startInfo);
       //检查重名
-      if (cml.options.has('-name')) {
-        var appName = cml.options.getValue('-name');
-        if (processLog.get(appName)) {
-          return console.warn("Application name already exists: " + appName);
-        }
+      if (processLog.get(params.name)) {
+        return console.warn("Application name already exists: " + params.name);
       }
       //请求启动
       notifier.waiting(1);
-      processMgr.startAppByInfo(startInfo);
-      break;
-    case "stop":
-      var nameOrPid = cml.args[0];
-      var isAll = !nameOrPid || nameOrPid == 'all';
-      if (!isAll && !processLog.get(nameOrPid)) {
+      processMgr.startApp(params);
+      return false;
+    }, false)
+
+    /**
+     * 停止一个应用
+     **/
+    .root.command('stop')
+    .action(function () {
+      var nameOrPid = this.argv[0];
+      if (nameOrPid && !processLog.get(nameOrPid)) {
         return console.warn('Can\'t find specified application: ' + nameOrPid);
       }
-      if (isAll) {
-        processMgr.killAllApp();
-        console.log("All application has stopped");
-      } else {
+      if (nameOrPid) {
         processMgr.killApp(nameOrPid);
         console.log("Stopped specified application: " + nameOrPid);
+      } else {
+        processMgr.killAllApp();
+        console.log("All application has stopped");
       }
-      break;
-    case "restart":
+      return false;
+    }, false)
+
+    /**
+     * 重启启用
+     **/
+    .root.command(['restart'])
+    .action(function () {
       var processCount = processLog.readArray().length;
       if (processCount < 1) {
         console.log("No application has been started");
         return;
       }
       console.log("Restarting...");
-      var nameOrPid = cml.args[0];
-      var isAll = !nameOrPid || nameOrPid == 'all';
-      if (!isAll && !processLog.get(nameOrPid)) {
+      var nameOrPid = this.argv[0];
+      if (nameOrPid && !processLog.get(nameOrPid)) {
         return console.warn('Can\'t find specified application: ' + nameOrPid);
       }
-      notifier.waiting(isAll ? processCount : 1);
-      if (isAll) {
-        processMgr.restartAllApp();
-      } else {
+      notifier.waiting(nameOrPid ? 1 : processCount);
+      if (nameOrPid) {
         processMgr.restartApp(nameOrPid);
+      } else {
+        processMgr.restartAllApp();
       }
-      break;
-    case "remove":
-    case "delete":
-      var nameOrPid = cml.args[0];
-      var isAll = !nameOrPid || nameOrPid == 'all';
-      if (!isAll && !processLog.get(nameOrPid)) {
+      return false;
+    }, false)
+
+    /**
+     * 停止并删除应用 
+     **/
+    .root.command(['remove', 'delete'])
+    .action(function () {
+      var nameOrPid = this.argv[0];
+      if (nameOrPid && !processLog.get(nameOrPid)) {
         return console.warn('Can\'t find specified application: ' + nameOrPid);
       }
-      if (isAll) {
-        processMgr.deleteAllApp();
-        console.log("All application has deleted");
-      } else {
+      if (nameOrPid) {
         processMgr.deleteApp(nameOrPid);
         console.log("Deleted specified application: " + nameOrPid);
+      } else {
+        processMgr.deleteAllApp();
+        console.log("All application has deleted");
       }
-      break;
-    case "ls":
-    case "list":
+      return false;
+    }, false)
+
+    /**
+     * 显示所有应用
+     **/
+    .root.command(['list', 'ls'])
+    .action(function () {
       var logArray = processLog.toPrintArray();
       if (logArray && logArray.length > 0) {
         console.log('Launched applications:' + env.EOL);
@@ -188,13 +166,24 @@ dm.run(function () {
       } else {
         console.log('No application has been started');
       }
-      break;
-    case "bootstrap":
-      if (cml.args[0]) {
-        var status = Boolean(cml.args[0]);
-        console.log(bootstrap.set(status));
-        return;
-      }
+      return false;
+    }, false)
+
+    /**
+     * 添加自启动
+     **/
+    .root.command(['startup'])
+    .action(function () {
+      var status = Boolean(this.argv[0] == 'on');
+      console.log(startup.set(status));
+      return false;
+    }, false)
+
+    /**
+     * 恢复意外终止的应用
+     **/
+    .root.command(['recovery'])
+    .action(function () {
       var appArray = processLog.readArray()
         .filter(function (app) {
           return app.status;
@@ -203,15 +192,22 @@ dm.run(function () {
         console.log("No application has been started");
         return;
       }
-      console.log("Starting...");
-      notifier.waiting(appArray.length);
-      processMgr.killAllApp();
-      appArray.forEach(function (app) {
-        processMgr.startApp(app);
-      })
-      break;
-    default:
-      console.error('Unrecognizable command "' + cml.command + '"');
-  }
+      console.log("Recovery...");
+      var nameOrPid = this.argv[0];
+      if (nameOrPid && !processLog.get(nameOrPid)) {
+        return console.warn('Can\'t find specified application: ' + nameOrPid);
+      }
+      notifier.waiting(nameOrPid ? 1 : appArray.length);
+      if (nameOrPid) {
+        processMgr.restartApp(nameOrPid);
+      } else {
+        appArray.forEach(function (log) {
+          processMgr.restartApp(log.name);
+        });
+      }
+    }, false)
+
+    //就续开始执行
+    .root.ready();
 
 });