+++ test/app/api/item.js
@@ -1,7 +1,7 @@
 /**
  * ItemController
  **/
-var ItemController = function() { };
+const ItemController = function() { };
 
 /**
  * 初始化方法，每次请求都会先执行 init 方法